code cleanup and fixing AgentViewController
